Changed view to menu. PlayerMenu handles responses.
