--- development_behavior_packs/portal_pal_BP/src/ViewManager.ts
@@ -1,3 +1,0 @@
-export class ViewManager {
-    
-}
